+++ src/log.c
@@ -8,7 +8,7 @@
 #define LOG_LEVEL LOG_INFO
 
 // Converts a log level int to a human readable string.
-static char* log_level_string(int level) {
+static char* log_level_string(uint8_t level) {
     switch (level) {
         case LOG_INFO:
             return "INFO";
@@ -31,7 +31,7 @@ void log_init() {
 
 // Prints a line with the log level to the log COM port
 // if the message level is at least LOG_LEVEL.
-void log_println(char const *msg, int level) {
+void log_println(char const *msg, uint8_t level) {
     if (LOG_LEVEL >= level) {
         serial_print_string(LOG_COM, log_level_string(level));
         serial_print_string(LOG_COM, ": ");
